project3_copilot/requirements.txt
openai==1.30.1
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
pydantic==2.7.1
